Fixed HTML errors
https://myparcelnl.github.io/api/#7_A_1
status -> Required: yes,
But there is no status in the example (https://myparcelnl.github.io/api/#7_A_2).
So I don't think it's required

https://myparcelnl.github.io/api/#7_B
Number has to be an integer, but when using a integer i get the following exception: Integer value found, but a string is required

https://myparcelnl.github.io/api/#6_C_4
Response says a statusCode 204 will be returned. But I receive a 201.

https://myparcelnl.github.io/api/#6_D_1
DownloadUrl object is not listed in the documentation

https://myparcelnl.github.io/api/#6_E_1
Shouldn't the parameters, Q, DROPOFF_TODAY, STATUS, FROM en TO, be Request parameters, instead of URI parameters?

https://myparcelnl.github.io/api/#7_A_1
Carrier is required it says.
But when calling GetShipment carrier is not returned.

https://myparcelnl.github.io/api/#7
TrackTrace object is not listed in the documentation.
https://myparcelnl.github.io/api/#6_G_4
property "Final" is also returned, but not listed in the documentation.

https://myparcelnl.github.io/api/#8_A
Object Definitions for DeliveryOptions and PickupOptions are not listed in the documentation.

https://myparcelnl.github.io/api/#9_A_2
Request body should be: array of Subscription objects.
Response body should ne: array of SubscriptionId objects.
Response is a 200 OK (instead of a 201 OK)

https://myparcelnl.github.io/api/#9_B
URI should be: https://api.myparcel.nl/webhook_subscriptions/ID/[;ID] (with ID URI parameter)
Authorization: basic BASE64(API_KEY OR SESSION_TOKEN) (also required?)
URI parameters ID = Subscription.id.
Response says a statusCode 201 is returned. But I receive a 204. (does not align with DeleteShipment)

https://myparcelnl.github.io/api/#9_C
Parameters description is missing

https://myparcelnl.github.io/api/#9_C_3
Id URI is required, so the example is not correct